fixed bug that caused an error when the wallpaper symlink doesn't exist

+++ random_wallpaper.py
@@ -1,19 +1,20 @@
 #!/usr/bin/python3
 
 from os import listdir, symlink, unlink
-from os.path import isfile, join
+from os.path import isfile, join, islink
 from subprocess import call
 from random import choice
 from shlex import quote
 
 wp_dir ="/media/green/Images/wallpapers"
-wp_img ="/home/gianluca/Immagini/wallpaper"
+wp_img ="/home/gianluca/Pictures/wallpaper"
 
 img_files = [f for f in listdir(wp_dir) if isfile(join(wp_dir, f))]
 
 selected_file = join(wp_dir, choice(img_files))
 escaped_file = quote(selected_file)
-unlink(wp_img)
+if islink(wp_img):
+    unlink(wp_img)
 symlink(escaped_file, wp_img)
 
 print(escaped_file)
